+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version changes are pinned to SDK releases.
 
 ## [Unreleased]
 
+- program: Add an ImmediateOrCancel order type. ([#168](https://github.com/zetamarkets/sdk/pull/168))
 - general: Add perpetual futures. (#[152](https://github.com/zetamarkets/sdk/pull/152)). Many changes, the main ones are:
   - New client.placePerpOrder() function
   - New ApplyFundingEvent